Added BitField class used to manipulated values of <8 bit stored contiguously in memory. Added MathUtils along with some helper functions Added FreeType FontBuilder Added FreeType 2.7.1 dependency Small improvements to AssertMsg() and added Assert() More functionnality to BitmapData (FlipX(), Copy, Convert in place now) BitmapData can now convert any format pair Replaced NULL with nullptr Added padding parameter to Texture Packer Full features TGA saver/loader (except palettes)

+++ Base/Profiler.h
@@ -24,7 +24,7 @@ private:
 	__int64 m_i64StartTime = 0;
 	__int64 m_i64StopTime = 0;
 
-	float m_fBlockTimeMs = 0.f;
+	double m_fBlockTimeMs = 0.f;
 
 	char* m_SectionID = nullptr;
