+++ OS_Base/Window.cpp
@@ -42,7 +42,7 @@ void Window::UnregisterMessageCallback(IWindowMessageCallback* pCallback)
 bool Window::NotifyCallbacks(void* hwnd, unsigned int message, uint64_t wParam, int64_t lParam)
 {
 	bool bHandled = false;
-	for (int i = 0; i < m_vMessageCallbacks.size(); i++)
+	for (size_t i = 0; i < m_vMessageCallbacks.size(); i++)
 	{
 		bHandled = m_vMessageCallbacks[i]->MessageCallback(hwnd, message, wParam, lParam);
 	}
